requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9
pydantic-settings>=2.4
sqlalchemy>=2.0
alembic>=1.13
asyncpg>=0.29
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
xgboost>=2.1
lightgbm>=4.5
mlflow>=2.16
redis>=5.0
httpx>=0.27
tenacity>=8.2  # 重试机制
bentoml>=1.2
python-dotenv>=1.0
loguru>=0.7
apache-airflow>=2.8  # 数据管道调度（可选，独立部署）
